requirements.txt
beautifulsoup4==4.6.0
mechanize
pymongo
python-dateutil
raven==6.9.0
requests
xlsxwriter